+++ app/src/main/java/org/simple/clinic/bloodsugar/BloodSugarMeasurement.kt
@@ -99,6 +99,12 @@ data class BloodSugarMeasurement(
       ORDER BY recordedAt DESC
     """)
     fun allBloodSugars(patientUuid: UUID): Observable<List<BloodSugarMeasurement>>
+    @Query("""
+      SELECT * FROM BloodSugarMeasurements
+      WHERE patientUuid == :patientUuid AND deletedAt IS NULL
+      ORDER BY recordedAt DESC
+    """)
+    fun allBloodSugarsImmediate(patientUuid: UUID): List<BloodSugarMeasurement>
 
     @Query("""
       SELECT * FROM BloodSugarMeasurements